fixing some conversion warnings

+++ src/memory.c
@@ -80,7 +80,7 @@ static inline unsigned int status(unsigned int i) {
   unsigned int word_ix = ix >> 5;    // / 32
   unsigned int bit_ix  = ix & 0x1F;  // % 32
 
-  uint32_t mask = 3 << bit_ix;       // 000110..0
+  uint32_t mask = ((uint32_t)3) << bit_ix;       // 000110..0
   return (bitmap[word_ix] & mask) >> bit_ix;
 }
 
@@ -89,7 +89,7 @@ static inline void set_status(unsigned int i, uint32_t status) {
   unsigned int word_ix = ix >> 5;    // / 32
   unsigned int bit_ix  = ix & 0x1F;  // % 32
 
-  uint32_t clr_mask = ~(3 << bit_ix);
+  uint32_t clr_mask = ~(((uint32_t)3) << bit_ix);
   uint32_t mask = status << bit_ix;
   bitmap[word_ix] &= clr_mask;
   bitmap[word_ix] |= mask;
